Updated Search Functionality
Added search routes which redirect to the relevant public gallery based
on the arguments in the header form.

Issues:

-We currently don’t have a good way to indicate that an assignment
number wasn’t found, so we currently redirect to the user’s own
gallery. We might need a separate page?
- The public gallery seems to be searching for assignments that are
flagged as ‘public: false’ which I thought was the opposite of what we
wanted.